update currentUserOrg set / get
