+++ components/ReferenceManager/references/reference_organizer/ReferenceProjectNavbarElOptions.tsx
@@ -8,7 +8,6 @@ import {
 import MoreVertIcon from "@mui/icons-material/MoreVert";
 import Menu from "@mui/material/Menu";
 import MenuItem from "@mui/material/MenuItem";
-import { useReferenceUploadDrawerContext } from "../reference_uploader/context/ReferenceUploadDrawerContext";
 
 type Props = {
   onSelectAddNewReference: (event: SyntheticEvent) => void;
